remove the unused 'ami' field from default AWS datacenter config
In the `AWSSpec` structure that describes an AWS datacenter, there was a
field named `AMI`. In most of our configurations it seems to have been set to
an old CoreOS image, and it appears to have been introduced back when
CoreOS was the only supported system. It hasn't been used since commit
`2f7589453edd74300513a78cdd971a9b23d7a0c6` (Henrik Schmidt,  Apr 4
08:47:44 2017) `Added support for node specific container linux version
(#298)`. We've removed the field in kubermatic/kubermatic@485a88c3 .